refactor: reorganize and fix compilation errors for no_std support
- Reorganize wrt-component into logical modules (async, canonical_abi, components, etc.)
- Add missing error codes (INVALID_BINARY, RESOURCE_EXHAUSTED, INVALID_ARGUMENT)
- Fix format! macro for no_std environments
- Add Duration import and feature-gated threading support
- Fix SafeStack imports and type mismatches
- Update instruction variant pattern matching
- Fix BoundedHashMap usage in no_std mode
- Correct Export field access (use .name and .ty instead of tuple access)
- Fix memory size access methods
- Add comprehensive imports for sync primitives across feature configurations
- Remove duplicate files and consolidate functionality

Still addressing compilation errors across std/no_std/alloc configurations.

+++ wrt/src/resource_nostd.rs
@@ -97,7 +97,7 @@ impl<P: MemoryProvider + Default + Clone + PartialEq + Eq> BoundedResourceTable<
         };
 
         // Create resource
-        let resource_id = ResourceId(self.resources.len() as u64);
+        let resource_id = ResourceId(self.resources.len() as u32);
         let resource = BoundedResource { id: resource_id, resource_type, name, is_dropped: false };
 
         // Add to table
